+++ src/client/components/Root/index.js
@@ -16,4 +16,5 @@ const Root = React.createClass({
   }
 })
 
+
 export default Root
