+++ src/client/config/routes.js
@@ -1,14 +1,16 @@
 import React from 'react'
-import { Route, Redirect } from 'react-router-dom'
-import { connect } from 'react-redux'
+import { Route } from 'react-router-dom'
+
+// commons
+import PrivateRoute from '../common/PrivateRoute'
 
 // react containers
 import Root from '../components/Root'
-import AsyncRoute from '../components/AsyncRoute/AsyncRoute'
+import AsyncRoute from '../common/AsyncRoute'
 
 const routes = () => (
   <Root>
-    <LimitedAccessRoute exact path='/home' component={props => (
+    <PrivateRoute exact path='/home' component={props => (
       <AsyncRoute
         props={props}
         loadingPromise={
@@ -33,35 +35,5 @@ const routes = () => (
     )} />
   </Root>
 )
-
-const PrivateRoute = (newProps) => {
-  return (
-    <Route path={newProps.path} render={props => {
-      console.log(newProps.authentication)
-      if (newProps.authentication.isAuthenticated) {
-        return (
-          React.createElement(newProps.component, props)
-        )
-      } else {
-        return (
-          <Redirect to={{ pathname: '/' }} />
-        )
-      }
-    }} />
-  )
-}
-
-const mapStateToProps = ({authReducer}) => {
-  console.log(authReducer)
-  return {
-    authentication: {
-      isAuthenticating: authReducer.isAuthenticating,
-      isAuthenticated: authReducer.isAuthenticated,
-      status: authReducer.status
-    }
-  }
-}
-
-let LimitedAccessRoute = connect(mapStateToProps)(PrivateRoute)
 
 export default routes
